requirements.txt
boto3==1.9.215
botocore==1.12.215
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.1.0
django-heroku==0.3.1
django-storages==1.7.1
djangorestframework==3.9.1
docutils==0.15.2
gunicorn==19.9.0
jmespath==0.9.4
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3
django-filters==0.2.1